requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

chromadb==0.4.22
numpy<2.0.0
openai>=1.52.0
google-generativeai>=0.3.0
anthropic>=0.18.0

Pillow==10.2.0
rembg==2.0.54
requests==2.31.0

boto3==1.34.0

python-dotenv==1.0.0
